requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
greenlet==3.0.3
idna==3.6
psycopg2==2.9.9
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
